requirements.txt
python-docx
deep-translator
langdetect
lxml

